update kotlin to 2.3.0
verinice-veo#1116

+++ src/main/kotlin/org/veo/accounts/keycloak/LicenseService.kt
@@ -83,9 +83,9 @@ class LicenseService(
 
     fun getInstalledLicense(): License = findInstalledLicense() ?: throw MissingLicenseException()
 
-    fun getLicensedTotalUsers(): Int = getInstalledLicense().totalUsers.toInt()
+    fun getLicensedTotalUsers(): Int = getInstalledLicense().totalUsers
 
-    fun getLicensedTotalClients(): Int = getInstalledLicense().totalClients.toInt()
+    fun getLicensedTotalClients(): Int = getInstalledLicense().totalClients
 
     private fun parseLicense(licenseString: String): License {
         val signers: Collection<SignerInformation>
